requirements.txt
# MedImagePrep - Dependencias para preprocesamiento de imágenes médicas
# Núcleo de procesamiento de imágenes
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0

# Visualización y gráficos
matplotlib>=3.7.0

# Soporte para formato médico DICOM
pydicom>=2.3.0

# Manejo de imágenes PNG/JPG
Pillow>=9.5.0

# Aplicación web interactiva
streamlit>=1.28.0

# Generación de reportes PDF
reportlab>=4.0.0

# Procesamiento de imágenes y análisis de textura
scikit-image>=0.21.0

# Machine learning para segmentación y clasificación
scikit-learn>=1.3.0
